+++ src/app/modules/search-parameter/search-parameter.component.ts
@@ -10,9 +10,11 @@ import {
 import { FhirBackendService } from '../../shared/fhir-backend/fhir-backend.service';
 import {
   encodeFhirSearchParameter,
-  escapeFhirSearchParameter
+  escapeFhirSearchParameter,
+  escapeStringForRegExp
 } from '../../shared/utils';
 import { SelectedObservationCodes } from '../../types/selected-observation-codes';
+import { LiveAnnouncer } from '@angular/cdk/a11y';
 
 /**
  * Component for editing one resource search parameter
@@ -70,7 +72,10 @@ export class SearchParameterComponent
     return this.CODETYPES.includes(this.selectedParameter.type);
   }
 
-  constructor(private fhirBackend: FhirBackendService) {
+  constructor(
+    private fhirBackend: FhirBackendService,
+    private liveAnnoncer: LiveAnnouncer
+  ) {
     super();
   }
 
@@ -92,6 +97,9 @@ export class SearchParameterComponent
       if (this.selectedParameter) {
         this.parameterValue.setValue(
           this.selectedParameter.type === 'boolean' ? 'true' : ''
+        );
+        this.liveAnnoncer.announce(
+          `Selected ${value}. One or more new fields have appeared.`
         );
         if (this.selectedParameter.valueSet) {
           this.parameterValues = this.definitions.valueSets[
@@ -115,12 +123,15 @@ export class SearchParameterComponent
     });
   }
 
+  /**
+   * Filters options list by matching value.
+   * It does a left match at word boundaries, example:
+   * 'variable name' will test true for 'va' or 'na', but not 'ri' or 'le'.
+   */
   private _filter(value: string, options: any[]): string[] {
-    const filterValue = value.toLowerCase();
-
-    return options.filter((option) =>
-      option.displayName.toLowerCase().includes(filterValue)
-    );
+    const reg = `\\b${escapeStringForRegExp(value)}`;
+    const regEx = new RegExp(reg, 'i');
+    return options.filter((option) => regEx.test(option.displayName));
   }
 
   /**